requirements.txt
Django==1.6.1
django-annoying
South >= 0.8.1
